Move task externalized extensions to TaskSaver
